+++ backend/canisters/identity/api/src/updates/c2c_set_user_ids.rs
@@ -1,13 +1,10 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
-use types::UserId;
+use types::{SuccessOnly, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub users: Vec<(Principal, Option<UserId>)>,
 }
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-}
+pub type Response = SuccessOnly;