+++ backend/canisters/notifications_index/impl/src/updates/remove_subscriptions_for_user.rs
@@ -1,7 +1,7 @@
 use crate::{RuntimeState, mutate_state};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
-use notifications_index_canister::remove_subscriptions_for_user::{Response::*, *};
+use notifications_index_canister::remove_subscriptions_for_user::*;
 use stable_memory_map::StableMemoryMap;
 
 #[update(msgpack = true)]
@@ -15,5 +15,5 @@ fn remove_subscriptions_for_user_impl(state: &mut RuntimeState) -> Response {
     if let Some(user_id) = state.data.principal_to_user_id_map.get(&caller) {
         state.remove_all_subscriptions(user_id, state.env.now());
     }
-    Success
+    Response::Success
 }