+++ backend/canisters/user/api/src/updates/configure_wallet.rs
@@ -2,6 +2,7 @@ use crate::WalletConfig;
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
+use types::SuccessOnly;
 
 #[ts_export(user, configure_wallet)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -9,8 +10,4 @@ pub struct Args {
     pub config: WalletConfig,
 }
 
-#[ts_export(user, configure_wallet)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-}
+pub type Response = SuccessOnly;