+++ backend/canisters/user/api/src/updates/mark_message_activity_feed_read.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::TimestampMillis;
+use types::{SuccessOnly, TimestampMillis};
 
 #[ts_export(user, mark_message_activity_feed_read)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -9,8 +9,4 @@ pub struct Args {
     pub read_up_to: TimestampMillis,
 }
 
-#[ts_export(user, mark_message_activity_feed_read)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-}
+pub type Response = SuccessOnly;