+++ backend/canisters/user/impl/src/updates/c2c_community_canister.rs
@@ -3,7 +3,7 @@ use crate::{RuntimeState, execute_update};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use user_canister::CommunityCanisterEvent;
-use user_canister::c2c_community_canister::{Response::*, *};
+use user_canister::c2c_community_canister::*;
 
 #[update(guard = "caller_is_known_community_canister", msgpack = true)]
 #[trace]
@@ -35,5 +35,5 @@ fn c2c_notify_community_canister_events_impl(args: Args, state: &mut RuntimeStat
         state.notify_user_index_of_chit(now);
     }
 
-    Success
+    Response::Success
 }