+++ backend/canisters/user/impl/src/updates/mark_read.rs
@@ -3,7 +3,7 @@ use crate::{RuntimeState, execute_update};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use constants::OPENCHAT_BOT_USER_ID;
-use user_canister::mark_read::{Response::*, *};
+use user_canister::mark_read::*;
 use user_canister::{MarkMessagesReadArgs, UserCanisterEvent};
 
 #[update(guard = "caller_is_owner", msgpack = true)]
@@ -57,5 +57,5 @@ fn mark_read_impl(args: Args, state: &mut RuntimeState) -> Response {
         }
     }
 
-    Success
+    Response::Success
 }