+++ backend/canisters/user_index/api/src/updates/register_external_achievement.rs
@@ -1,7 +1,7 @@
 use candid::{CandidType, Principal};
 use human_readable::{HumanReadablePrincipal, ToHumanReadable};
 use serde::{Deserialize, Serialize};
-use types::{CanisterId, TimestampMillis, UserId};
+use types::{CanisterId, SuccessOnly, TimestampMillis, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -16,10 +16,7 @@ pub struct Args {
     pub max_awards: u32,
 }
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-}
+pub type Response = SuccessOnly;
 
 #[derive(Serialize)]
 pub struct HumanReadableArgs {