+++ backend/canisters/user_index/api/src/updates/remove_sms_messages.rs
@@ -1,12 +1,10 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
+use types::SuccessOnly;
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub up_to_index: u64,
 }
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-}
+pub type Response = SuccessOnly;