+++ backend/canisters/user_index/impl/src/updates/c2c_suspend_users.rs
@@ -3,7 +3,7 @@ use crate::timer_job_types::{SetUserSuspended, TimerJob};
 use crate::{RuntimeState, mutate_state};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
-use user_index_canister::c2c_suspend_users::{Response::*, *};
+use user_index_canister::c2c_suspend_users::*;
 
 #[update(guard = "caller_is_group_index", msgpack = true)]
 #[trace]
@@ -27,5 +27,5 @@ fn c2c_suspend_users_impl(args: Args, state: &mut RuntimeState) -> Response {
             );
         }
     }
-    Success
+    Response::Success
 }